requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# Google Cloud Platform
google-cloud-functions==1.16.0
google-cloud-run==0.10.0
google-cloud-secret-manager==2.18.0
google-cloud-bigquery==3.15.0
google-cloud-storage==2.16.0
google-cloud-pubsub==2.23.0
google-auth==2.27.0

# Neo4j Graph Database
neo4j==5.16.0
py2neo==2021.2.4

# Data Processing
pandas==2.2.0
numpy==1.26.3
jsonschema==4.21.0

# Simulation & Analysis
networkx==3.2.1
matplotlib>=3.9.0
seaborn==0.13.1

# API & Web (optional, for future API endpoint)
flask==3.0.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.1.1
flake8==7.0.0
pylint==3.0.3

# Jupyter (for analysis notebooks)
jupyter==1.0.0
ipython==8.20.0
jupyterlab==4.0.11

# Logging & Monitoring
structlog==24.1.0
colorlog==6.8.0

